"Registro, login y filtros"

+++ src/main/java/com/uade/tpo/cars_e_commerce/entity/ShopCart.java
@@ -20,7 +20,7 @@ public class ShopCart {
         private Long shopCartid;
 
         @Column
-        private long total;
+        private Double total;
 
         @OneToOne
         @JoinColumn(name = "user_id", nullable = false)
